refactor: ♻️ apply assorted ruff lint fixes (RUF005, RUF046, PLR1730, RET504, G004, DTZ011, D205)
- select.py: use unpacking `[x, *list]` instead of list concatenation (RUF005, 4 sites)

- number.py: drop redundant `int(round(raw))` cast (RUF046)

- coordinator.py: collapse `if x > max: x = max` into `min()` call (PLR1730)

- entity.py: return device_info dict directly without intermediate variable (RET504)

- helpers.py: same for get_device_time return path; reflow prepare_device_time docstring (D205, RET504)

- migration.py: reflow async_cleanup_legacy_files docstring (D205)

- select.py: switch backwash log statement from f-string to %-style placeholders (G004)

- options_flow.py: replace `date.today()` with `dt_util.now()` so the year reflects HA's configured timezone, not the host clock (DTZ011)

- update test_options_flow patches to mock `dt_util` instead of the dropped `date` import

+++ custom_components/neopool/number.py
@@ -273,7 +273,7 @@ class NeoPoolNumber(NeoPoolEntity, NumberEntity):  # type: ignore[reportIncompat
         if (
             self.suggested_display_precision == 0 and raw is not None
         ):  # pragma: no cover
-            return int(round(raw))
+            return round(raw)
         if raw is None:
             return self._attr_native_value  # fallback if coordinator is not updated yet
         return round(raw, 2)
